pre-gui preparations
