+++ lib/game-ids.json
@@ -1,0 +1,56 @@
+{
+  "1E95E5926F1CB99A87326D927F27B47E": "Nintendo Switch (controller screen)",
+  "CCFA659F4857F96DDA29AFEDB2E166E6": "Nintendo Switch (friend screen)",
+  "57B4628D2267231D57E0FC1078C0596D": "Nintendo Switch (main menu and news)",
+  "1AB131B6E6571375B79964211BB3F5AE": "Nintendo Switch (points screen)",
+
+  "02CB906EA538A35643C1E1484C4B947D": "Animal Crossing: New Horizons",
+  "EAD9CDD267F449963C89657382E411AE": "Baba Is You",
+  "13BBCD125931ACA6E654AE761A3B927E": "Broforce",
+  "0113217627AA97E2DF5A859696D2A2C1": "Darkest Dungeon",
+  "3D69A7ED02A1FF371048829E22A49194": "Diablo III: Eternal Collection",
+  "E58C03FD2FF562F462515FB73597EDA6": "DRAGON QUEST XI S: Echoes of an Elusive Age – Definitive Edition",
+  "DC083B2CD4ED80E7BC9C8AD7A279BB45": "DRAGON QUEST XI S: Echoes of an Elusive Age – Definitive Edition [Demo version]",
+  "7F2F1111AFE3AA11E958B0ACD0B5B925": "Enter the Gungeon",
+  "07ACA5809C9608779ED2B95B8E58398F": "FRAMED COLLECTION",
+  "EF362055E2612FF3E04AF0394A9E0FE9": "Hiragana Pixel Party",
+  "5D315FFC05DFD3E4307A98D4E0AE3ED4": "Human: Fall Flat",
+  "86C14151EAC920165D017C83A8A5AAE9": "Kingdom: New Lands",
+  "FB761F89764FE6715E1C3E00FCB812D3": "LEGO Harry Potter Collection",
+  "73F83DA721DE5AFEED1F917A300ED622": "MagiCat",
+  "16851BE00BC6068871FE49D98876D6C5": "Mario Kart 8 Deluxe",
+  "BA8E0FCB931BBA7F76DCA06732EF9A07": "Mario Tennis Aces",
+  "2EC14E01DAA280E5A2B7C4BA9BF1289D": "Mini Metro",
+  "5310A75890314E1C83A6E33900C17392": "Nights of Azure 2: Bride of the New Moon",
+  "8F655652CF5441D7471D936F3F07324D": "Nintendo Entertainment System - Nintendo Switch Online",
+  "F897DAA49269BC023482B5C9AAA6BE73": "Overcooked! Special Edition",
+  "86EB443FBFF5CB1932FDC277D68F2DF5": "Peace, Death!",
+  "9E2A5260CC77462B720A866290153029": "Project Highrise: Architect's Edition",
+  "0585E865DFB68B5298F19360A730EDB3": "Puyo Puyo Tetris",
+  "638E7E1EEC4CD8A239243633C0345A07": "Ring Fit Adventure",
+  "5D837F322A5F885403B2440059A0308C": "Risk of Rain",
+  "E8680D795743293DE312E20C9B85C1DD": "Risk of Rain 2",
+  "64E0CC877FA4951A225029BBE709009F": "Robonauts",
+  "4947BC16E9EFC6B6E5686953E6F8B74C": "Runbow",
+  "42A65F67C20F2AF4737AE9D4D5A37ECB": "Slay the Spire",
+  "66C396DFF5BBB36B52AEA44F72B55406": "Snipperclips Plus",
+  "1628E0CE3F839127054B0EE36E28E52A": "Sonic Mania",
+  "397A963DA4660090D65D330174AC6B04": "Splatoon 2",
+  "66B80398A8C97BF0A148FEF3C6E256BA": "Star Wars Episode I: Racer",
+  "8E4B077462389FB7FC569E446EFD8291": "Star Wars Pinball",
+  "4D5A2A79927AEFDD7EBB9B111EF09EA1": "SteamWorld Dig",
+  "96794F2558A5B75BC0DBCC69AA0119EE": "Super Bomberman R",
+  "502BADF490303E7A09F1BF58AD4341AB": "Super Mario Bros. 35",
+  "BF19FBEA37724338D87F26F17A3B97B2": "Super Mario Maker 2",
+  "8AEDFF741E2D23FBED39474178692DAF": "Super Mario Odyssey",
+  "099ECEEF904DB62AEE3A76A3137C241B": "Super Mario Party",
+  "05DC14F80A13996B94160CD375AFD506": "Super Nintendo Entertainment System - Nintendo Switch Online",
+  "0E7DF678130F4F0FA2C88AE72B47AFDF": "Super Smash Bros. Ultimate",
+  "DB82B254F554D50F19AFC1EBEE22498E": "Tardy",
+  "691C9B2C6D1F1E032DDC01FD026159FD": "TETRIS 99",
+  "1B7686315C6209EBBD25E3E11E89316C": "The Binding of Isaac: Afterbirth+",
+  "463C9DA3456CB06F891D460282D6F4AC": "The Bridge",
+  "F1C11A22FAEE3B82F21B330E1B786A39": "The Legend of Zelda: Breath of the Wild",
+  "47645B79AAC91BAA27FBE1B4F9080550": "Where Are My Friends?",
+  "7AEA3B76283DF2B97E581259A12F733D": "Yoshi's Crafted World [Demo]"
+}
