+++ lib/logger.js
@@ -1,0 +1,7 @@
+function print(text) {
+  if(global.args.workMode === 'cmd') {
+    process.stdout.write(`${text}\n`);
+  }
+}
+
+module.exports = print;
