+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nsw-screenshot-renamer",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -14,5 +14,8 @@
   },
   "dependencies": {
     "commander": "^7.2.0"
+  },
+  "devDependencies": {
+    "electron": "^12.0.2"
   }
 }
